Change serializer for test runs
